refactor(remote): merge modules, strip password handling, improve health checks
Problem:
Review feedback identified several issues: private API docs exposed
publicly, password prompt handling added unnecessary complexity, health
checks, used manual /tmp validation instead of ssh-native checks,
version parsing was done manually instead of using vim.version APIs.

Solution:
- Merge _remote.lua into _ssh.lua (group by topic).
- Remove all password/headless/PTY codepath (SSH key-only auth for now).
- Replace manual /tmp check with `ssh -G` config dump.
- Use vim.version.parse() for SSH version checking and vim.version() for
Nvim version.
- Use stdpath('run') for mux socket path.
- Remove private Lua API docs section.
- Fix docs wording per review.
- Add E6101 error code with help tag.
- Add news.txt entry.
