+++ src/nvim/ex_docmd.c
@@ -6011,17 +6011,17 @@ static void ex_detach(exarg_T *eap)
 /// ":connect! <address>" stops the current server if no other UIs are attached, then connects to the given server.
 static void ex_connect(exarg_T *eap)
 {
-  char *address = eap->arg;
+  char *addr = eap->arg;
   typval_T ssh_rettv = { .v_type = VAR_UNKNOWN };
 
   // SSH URI: bootstrap remote Neovim, get local forwarded socket
-  if (strncmp(address, "ssh://", 6) == 0) {
+  if (strncmp(addr, "ssh://", 6) == 0) {
     typval_T argvars[2];
     argvars[0].v_type = VAR_STRING;
-    argvars[0].vval.v_string = address;
+    argvars[0].vval.v_string = addr;
     argvars[1].v_type = VAR_UNKNOWN;
 
-    nlua_call_typval("vim.net._remote", "start", argvars, &ssh_rettv);
+    nlua_call_typval("vim.net._ssh", "start", argvars, &ssh_rettv);
 
     if (ssh_rettv.v_type != VAR_STRING || ssh_rettv.vval.v_string == NULL) {
       emsg("E6101: SSH connection failed");
@@ -6029,13 +6029,13 @@ static void ex_connect(exarg_T *eap)
       return;
     }
 
-    address = ssh_rettv.vval.v_string;
+    addr = ssh_rettv.vval.v_string;
   }
 
   bool stop_server = eap->forceit ? (ui_active() == 1) : false;
 
   Error err = ERROR_INIT;
-  remote_ui_connect(current_ui, address, &err);
+  remote_ui_connect(current_ui, addr, &err);
   tv_clear(&ssh_rettv);
 
   if (ERROR_SET(&err)) {
