+++ test/functional/lua/ssh_spec.lua
@@ -106,7 +106,7 @@ describe('SSH parser', function()
   end)
 end)
 
-describe('vim.net._remote', function()
+describe('vim.net._ssh', function()
   local fake_bin_dir
 
   local function setup_fake_ssh(behavior)
@@ -119,117 +119,105 @@ describe('vim.net._remote', function()
 
     local script
     if is_windows then
-      script = [=[
-@echo off
-setlocal EnableDelayedExpansion
-set "ARGS=%*"
+      script = t.dedent([=[
+        @echo off
+        setlocal EnableDelayedExpansion
+        set "ARGS=%*"
 
-echo(!ARGS! | findstr /C:"uname -s && uname -m" >nul
-if not errorlevel 1 (
-]=] .. (behavior.uname or [=[
-  echo Linux
-  echo x86_64
-]=]) .. [=[
-  exit /b 0
-)
+        echo(!ARGS! | findstr /C:"uname -s && uname -m" >nul
+        if not errorlevel 1 (
+      ]=] .. (behavior.uname or [=[
+          echo Linux
+          echo x86_64
+      ]=]) .. [=[
+          exit /b 0
+        )
 
-echo(!ARGS! | findstr /C:"-L" >nul
-if not errorlevel 1 (
-  echo(!ARGS! | findstr /C:"ControlMaster" >nul
-  if errorlevel 1 (
-    echo FAIL: Multiplexing flags missing! 1>&2
-    exit /b 1
-  )
-]=] .. (behavior.tunnel or [=[
-  echo Password: 1>&2
-  set /p PASS=
-  if not "!PASS!"=="secret" (
-    echo Access denied 1>&2
-    exit /b 1
-  )
-  set "SOCK="
-  set "NEXT_L="
-  for %%A in (%*) do (
-    if defined NEXT_L (
-      set "SOCK=%%~A"
-      set "NEXT_L="
-    ) else if "%%~A"=="-L" (
-      set "NEXT_L=1"
-    )
-  )
-  if not defined SOCK (
-    echo Unexpected SSH command: !ARGS! 1>&2
-    exit /b 1
-  )
-  > "!SOCK!" type NUL
-  exit /b 0
-]=]) .. [=[
-)
+        echo(!ARGS! | findstr /C:"-L" >nul
+        if not errorlevel 1 (
+          echo(!ARGS! | findstr /C:"ControlMaster" >nul
+          if errorlevel 1 (
+            echo FAIL: Multiplexing flags missing! 1>&2
+            exit /b 1
+          )
+      ]=] .. (behavior.tunnel or [=[
+          set "SOCK="
+          set "NEXT_L="
+          for %%A in (%*) do (
+            if defined NEXT_L (
+              set "SOCK=%%~A"
+              set "NEXT_L="
+            ) else if "%%~A"=="-L" (
+              set "NEXT_L=1"
+            )
+          )
+          if not defined SOCK (
+            echo Unexpected SSH command: !ARGS! 1>&2
+            exit /b 1
+          )
+          > "!SOCK!" type NUL
+          exit /b 0
+      ]=]) .. [=[
+        )
 
-echo(!ARGS! | findstr /C:"TARGET_VER" >nul
-if not errorlevel 1 (
-]=] .. (behavior.installer or [=[
-  echo Installing Neovim... 1>&2
-  exit /b 0
-]=]) .. [=[
-)
+        echo(!ARGS! | findstr /C:"TARGET_VER" >nul
+        if not errorlevel 1 (
+      ]=] .. (behavior.installer or [=[
+          echo Installing Neovim... 1>&2
+          exit /b 0
+      ]=]) .. [=[
+        )
 
-echo(!ARGS! | findstr /C:"-O" >nul
-if not errorlevel 1 (
-  echo(!ARGS! | findstr /C:"exit" >nul
-  if not errorlevel 1 exit /b 0
-)
+        echo(!ARGS! | findstr /C:"-O" >nul
+        if not errorlevel 1 (
+          echo(!ARGS! | findstr /C:"exit" >nul
+          if not errorlevel 1 exit /b 0
+        )
 
-echo Unexpected SSH command: !ARGS! 1>&2
-exit /b 1
-]=]
+        echo Unexpected SSH command: !ARGS! 1>&2
+        exit /b 1
+      ]=])
     else
-      script = [=[
-#!/usr/bin/env bash
-ARGS="$*"
+      script = t.dedent([=[
+        #!/usr/bin/env bash
+        ARGS="$*"
 
-if [[ "$ARGS" == *"uname -s && uname -m"* ]]; then
-]=] .. (behavior.uname or [=[
-  echo "Linux"
-  echo "x86_64"
-]=]) .. [=[
-  exit 0
-fi
+        if [[ "$ARGS" == *"uname -s && uname -m"* ]]; then
+      ]=] .. (behavior.uname or [=[
+          echo "Linux"
+          echo "x86_64"
+      ]=]) .. [=[
+          exit 0
+        fi
 
-if [[ "$ARGS" == *"-L"* ]]; then
-  if [[ "$ARGS" != *"ControlMaster"* ]]; then
-    echo "FAIL: Multiplexing flags missing!" >&2
-    exit 1
-  fi
-]=] .. (behavior.tunnel or [=[
-  echo "Password: " >&2
-  read -r PASS
-  PASS=$(echo "$PASS" | tr -d '\r')
-  if [ "$PASS" != "secret" ]; then
-    echo "Access denied" >&2
-    exit 1
-  fi
-  SOCK=$(echo "$ARGS" | grep -oE '\-L [^:]+' | cut -d' ' -f2)
-  touch "$SOCK"
-  sleep 60 &
-  exit 0
-]=]) .. [=[
-fi
+        if [[ "$ARGS" == *"-L"* ]]; then
+          if [[ "$ARGS" != *"ControlMaster"* ]]; then
+            echo "FAIL: Multiplexing flags missing!" >&2
+            exit 1
+          fi
+      ]=] .. (behavior.tunnel or [=[
+          SOCK=$(echo "$ARGS" | grep -oE '\-L [^:]+' | cut -d' ' -f2)
+          echo "NVIM_READY"
+          touch "$SOCK"
+          sleep 60 &
+          exit 0
+      ]=]) .. [=[
+        fi
 
-if [[ "$ARGS" == *"TARGET_VER"* ]]; then
-]=] .. (behavior.installer or [=[
-  echo "Installing Neovim..." >&2
-  exit 0
-]=]) .. [=[
-fi
+        if [[ "$ARGS" == *"TARGET_VER"* ]]; then
+      ]=] .. (behavior.installer or [=[
+          echo "Installing Neovim..." >&2
+          exit 0
+      ]=]) .. [=[
+        fi
 
-if [[ "$ARGS" == *"-O"*"exit"* ]]; then
-  exit 0
-fi
+        if [[ "$ARGS" == *"-O"*"exit"* ]]; then
+          exit 0
+        fi
 
-echo "Unexpected SSH command: $ARGS" >&2
-exit 1
-]=]
+        echo "Unexpected SSH command: $ARGS" >&2
+        exit 1
+      ]=])
     end
     t.write_file(fake_ssh_path, script)
     vim.uv.fs_chmod(fake_ssh_path, 493)
@@ -239,7 +227,6 @@ exit 1
       [[
       _G.orig_path = vim.fn.getenv('PATH')
       vim.fn.setenv('PATH', %q .. %q .. _G.orig_path)
-      vim.fn.setenv('NVIM_TEST_MOCK_UI', '1')
     ]],
       fake_bin_dir,
       path_sep
@@ -274,7 +261,7 @@ exit 1
         ]],
       })
       local res = n.exec_lua([[
-        return { require('vim.net._remote').get_system_info({ host = 'server' }) }
+        return { require('vim.net._ssh').get_system_info({ host = 'server' }) }
       ]])
       eq('linux', res[1])
       eq('x86_64', res[2])
@@ -289,7 +276,7 @@ exit 1
         ]],
       })
       local res = n.exec_lua([[
-        return { require('vim.net._remote').get_system_info({ host = 'server' }) }
+        return { require('vim.net._ssh').get_system_info({ host = 'server' }) }
       ]])
       eq('darwin', res[1])
       eq('arm64', res[2])
@@ -305,7 +292,7 @@ exit 1
       })
       local status, err = pcall(function()
         n.exec_lua([[
-          require('vim.net._remote').get_system_info({ host = 'server' })
+          require('vim.net._ssh').get_system_info({ host = 'server' })
         ]])
       end)
       eq(false, status)
@@ -314,24 +301,14 @@ exit 1
   end)
 
   describe('Remote Engine (Orchestration)', function()
-    it('injects password and returns local socket', function()
+    it('returns local socket with key-based auth', function()
       skip(is_os('win'), 'remote-ssh engine is POSIX-only')
       setup_fake_ssh()
-      local res = n.exec_lua([[
-        _G.inputs_requested = {}
-        vim.ui.input = function(opts, on_confirm)
-          table.insert(_G.inputs_requested, opts.prompt)
-          on_confirm("secret")
-        end
-        local sock = require('vim.net._remote').start('user@test-server')
-        return { sock, _G.inputs_requested }
+      local sock = n.exec_lua([[
+        return require('vim.net._ssh').start('user@test-server')
       ]])
 
-      local sock = res[1]
-      local inputs = res[2]
-
       assert(sock:match('_remote_nvim%.sock$'))
-      eq('Password: ', inputs[1])
     end)
   end)
 end)
